agents/backend/onyx/server/features/blaze_ai/requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
orjson>=3.9.0

# AI/ML Core Libraries
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.36.0
diffusers>=0.24.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
safetensors>=0.4.0
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Advanced AI Libraries
xformers>=0.0.22
flash-attn>=2.3.0
optimum>=1.15.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.15.0
evaluate>=0.4.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Text Processing & NLP
nltk>=3.8.0
textstat>=0.7.0
yake>=0.4.8
spacy>=3.7.0
textblob>=0.17.0
gensim>=4.3.0
wordcloud>=1.9.0

# Data Science & Analytics
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities & Performance
psutil>=5.9.0
pyyaml>=6.0.0
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.0
asyncio-mqtt>=0.16.0
redis>=5.0.0
celery>=5.3.0

# Monitoring & Observability
prometheus-client>=0.19.0
elasticsearch>=8.11.0
jaeger-client>=4.8.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Web & API
gradio>=4.0.0
streamlit>=1.28.0
dash>=2.14.0
flask>=3.0.0
django>=4.2.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
motor>=3.3.0
pymongo>=4.6.0

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0
tensorboard>=2.14.0
wandb>=0.16.0
mlflow>=2.8.0
dvc>=3.30.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0

# Cloud & Deployment
boto3>=1.34.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.19.0
kubernetes>=28.1.0
docker>=6.1.0

# Optional: GPU Optimization
# nvidia-ml-py3>=7.352.0
# cupy-cuda12x>=12.0.0
# cudf-cu12>=23.8.0
# cuml-cu12>=23.8.0

# Optional: Advanced Features
# ray>=2.8.0
# dask>=2023.11.0
# vaex>=4.17.0
# polars>=0.19.0
# duckdb>=0.9.0

